+++ integrations/scala/RELEASE_NOTES.md
@@ -1,8 +1,8 @@
-# AngularTS Scala.js Facades 0.34.0
+# AngularTS Scala.js Facades 0.35.0
 
 Compatible runtime:
 
-- AngularTS npm package: 0.34.0
+- AngularTS npm package: 0.35.0
 - Maven coordinate: io.github.angular-wave:angular-ts-scala_sjs1_3
 
 ## Scope